chapter_4/requirements.txt
openai
python-dotenv
semantic-kernel